+++ refinery/tests/postgres.rs
@@ -615,7 +615,6 @@ mod postgres {
                     "migrate",
                     "-c",
                     "tests/postgres_refinery.toml",
-                    "files",
                     "-p",
                     "tests/migrations",
                 ])