+++ refinery/tests/tiberius.rs
@@ -898,7 +898,6 @@ mod tiberius {
                     "migrate",
                     "-c",
                     "tests/tiberius_refinery.toml",
-                    "files",
                     "-p",
                     "tests/migrations",
                 ])